requirements_streamlit.txt
streamlit==1.28.1
openai==1.97.1
python-dotenv==1.0.0 